+++ lib/helpers/component_helpers.ex
@@ -55,7 +55,10 @@ defmodule PrimerLive.Helpers.ComponentHelpers do
     end
   end
 
-  def missing_attribute(name) do
-    IO.puts("Missing attribute: #{name}")
+  def missing_attribute(component, name, info \\ "") do
+    message =
+      ["Missing attribute: '#{name}' for component '#{component}'.", info] |> Enum.join(" ")
+
+    IO.puts(message)
   end
 end